requirements.txt
python-telegram-bot==20.7
openai==1.12.0
aiosqlite==0.20.0
requests==2.31.0
beautifulsoup4==4.12.3
